# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CreateWorkoutSessionInputTrainingGoal = typeof CreateWorkoutSessionInputTrainingGoal[keyof typeof CreateWorkoutSessionInputTrainingGoal];


export const CreateWorkoutSessionInputTrainingGoal = {
  muscle_tone: 'muscle_tone',
  posture: 'posture',
  cardio_general: 'cardio_general',
  weight_loss: 'weight_loss',
} as const;

/**
 * @nullable
 */
export type CreateWorkoutSessionInputHealthSource = typeof CreateWorkoutSessionInputHealthSource[keyof typeof CreateWorkoutSessionInputHealthSource] | null;


export const CreateWorkoutSessionInputHealthSource = {
  health_connect: 'health_connect',
  healthkit: 'healthkit',
} as const;

export interface CreateWorkoutSessionInput {
  trainingGoal: CreateWorkoutSessionInputTrainingGoal;
  /** @minimum 0 */
  durationSeconds: number;
  isValid: boolean;
  /** @nullable */
  healthSource?: CreateWorkoutSessionInputHealthSource;
}

export interface WorkoutSessionRecord {
  id: number;
  createdAt: string;
}

/**
 * @nullable
 */
export type PricingStateSubscriptionStatus = typeof PricingStateSubscriptionStatus[keyof typeof PricingStateSubscriptionStatus] | null;


export const PricingStateSubscriptionStatus = {
  active: 'active',
  canceled: 'canceled',
  past_due: 'past_due',
  incomplete: 'incomplete',
} as const;

/**
 * currentLevel is 0-10 (zero-based). Level 0 = 250 EUR (≤8 sessions/month), Level 10 = 10 EUR (≥18 sessions/month). displayLevel = currentLevel + 1. Price is recalculated from scratch every month — no cycle counter.
 */
export interface PricingState {
  userId: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  currentLevel: number;
  /**
     * @minimum 1
     * @maximum 11
     */
  displayLevel: number;
  priceEur: number;
  /** @nullable */
  subscriptionStartedAt: string | null;
  lastMonthCompleted: boolean;
  /** @nullable */
  subscriptionStatus: PricingStateSubscriptionStatus;
}

export interface PricingCheckoutSession {
  url: string;
}

export interface PricingPortalSession {
  url: string;
}

export interface PricingCycleResult {
  processedCount: number;
  improvedCount: number;
  regressedCount: number;
  unchangedCount: number;
}