# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pricing
    description: 19 Sessions pricing state machine and Stripe billing
  - name: sessions
    description: Workout session recording
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /sessions:
    post:
      operationId: createWorkoutSession
      tags: [sessions]
      summary: Save a completed workout session
      description: >-
        Saves a completed workout session for the authenticated user. The
        isValid flag is determined client-side from Health Connect (Android)
        or HealthKit (iOS) — the server trusts and persists it as-is.
        userId is derived from the Clerk session, never from the request body.
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutSessionInput"
      responses:
        "201":
          description: Session saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutSessionRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pricing/state:
    get:
      operationId: getPricingState
      tags: [pricing]
      summary: Get pricing state for the authenticated user
      description: >-
        Returns the current pricing level and monthly progress for the
        authenticated caller (identified from the Clerk session, never from a
        client-supplied identifier). Creates a default Level 1 (currentLevel 0,
        250 EUR) row if none exists yet.
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Pricing state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingState"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pricing/checkout:
    post:
      operationId: createPricingCheckout
      tags: [pricing]
      summary: Create a Stripe Checkout session for the current level's price
      description: >-
        Operates on the authenticated caller only (Clerk session). The Stripe
        customer email is taken from the authenticated user's Clerk profile.
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingCheckoutSession"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pricing/portal:
    post:
      operationId: createPricingPortalSession
      tags: [pricing]
      summary: Create a Stripe customer portal session (cancel/manage billing)
      description: >-
        Operates on the authenticated caller only (Clerk session) — never on a
        userId supplied by the client.
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Portal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingPortalSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No Stripe customer for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pricing/cycle-all:
    post:
      operationId: runPricingCycle
      tags: [pricing]
      summary: Internal — run the monthly price recalculation for all users
      description: >-
        Cron-only endpoint, not called by the mobile app. Must be invoked once a
        month by a scheduled job (e.g. Replit Scheduled Deployment) with the
        shared secret header. Evaluates each user's previous calendar month
        (valid session count only — no reliability score), sets the new price
        via direct lookup, and syncs it to the user's active Stripe subscription.
        No cycle counter or gradual level transitions — each month is independent.
      parameters:
        - name: x-cron-secret
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cycle run summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingCycleResult"
        "401":
          description: Missing or invalid cron secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreateWorkoutSessionInput:
      type: object
      properties:
        trainingGoal:
          type: string
          enum: [muscle_tone, posture, cardio_general, weight_loss]
        durationSeconds:
          type: integer
          minimum: 0
        isValid:
          type: boolean
        healthSource:
          type: ["string", "null"]
          enum: ["health_connect", "healthkit", null]
      required:
        - trainingGoal
        - durationSeconds
        - isValid
    WorkoutSessionRecord:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - createdAt
    PricingState:
      type: object
      description: >-
        currentLevel is 0-10 (zero-based). Level 0 = 250 EUR (≤8 sessions/month),
        Level 10 = 10 EUR (≥18 sessions/month). displayLevel = currentLevel + 1.
        Price is recalculated from scratch every month — no cycle counter.
      properties:
        userId:
          type: string
        currentLevel:
          type: integer
          minimum: 0
          maximum: 10
        displayLevel:
          type: integer
          minimum: 1
          maximum: 11
        priceEur:
          type: integer
        subscriptionStartedAt:
          type: ["string", "null"]
          format: date-time
        lastMonthCompleted:
          type: boolean
        subscriptionStatus:
          type: ["string", "null"]
          enum: ["active", "canceled", "past_due", "incomplete", null]
      required:
        - userId
        - currentLevel
        - displayLevel
        - priceEur
        - subscriptionStartedAt
        - lastMonthCompleted
        - subscriptionStatus
    PricingCheckoutSession:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    PricingPortalSession:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    PricingCycleResult:
      type: object
      properties:
        processedCount:
          type: integer
        improvedCount:
          type: integer
        regressedCount:
          type: integer
        unchangedCount:
          type: integer
      required:
        - processedCount
        - improvedCount
        - regressedCount
        - unchangedCount
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      description: >-
        Clerk session token. Web clients send the session cookie
        automatically; mobile clients attach it as a Bearer token. The server
        resolves the authenticated userId from this token — it is never
        accepted from the request body/path/params.