# lib/api-zod/src/generated/types/createWorkoutSessionInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateWorkoutSessionInputHealthSource } from './createWorkoutSessionInputHealthSource';
import type { CreateWorkoutSessionInputTrainingGoal } from './createWorkoutSessionInputTrainingGoal';

export interface CreateWorkoutSessionInput {
  trainingGoal: CreateWorkoutSessionInputTrainingGoal;
  /** @minimum 0 */
  durationSeconds: number;
  isValid: boolean;
  /** @nullable */
  healthSource?: CreateWorkoutSessionInputHealthSource;
}