# lib/api-zod/src/generated/types/pricingState.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PricingStateSubscriptionStatus } from './pricingStateSubscriptionStatus';

/**
 * currentLevel is 0-10 (zero-based). Level 0 = 250 EUR (≤8 sessions/month), Level 10 = 10 EUR (≥18 sessions/month). displayLevel = currentLevel + 1. Price is recalculated from scratch every month — no cycle counter.
 */
export interface PricingState {
  userId: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  currentLevel: number;
  /**
     * @minimum 1
     * @maximum 11
     */
  displayLevel: number;
  priceEur: number;
  /** @nullable */
  subscriptionStartedAt: Date | null;
  lastMonthCompleted: boolean;
  /** @nullable */
  subscriptionStatus: PricingStateSubscriptionStatus;
}